+++ src/actions/commit.ts
@@ -6,13 +6,12 @@ import {
   validateStore,
   updateDataStoreMetadata,
   getLatestStoreInfo,
-  findStoreId,
 } from "../blockchain/datastore";
-import { serializeStoreInfo } from "../blockchain/serialization";
 import {
   DIG_FOLDER_PATH,
   getManifestFilePath,
   loadDigConfig,
+  getActiveStoreId,
 } from "../utils/config";
 import { waitForConfirmation } from "../blockchain/coins";
 import { getPeer } from "../blockchain/peer";
@@ -30,7 +29,8 @@ export const commit = async (): Promise<void> => {
       throw new Error("Store integrity check failed.");
     }
 
-    const storeId = findStoreId();
+    const storeId = await getActiveStoreId();
+
     if (!storeId) {
       throw new Error("Store ID not found. Please run init first.");
     }
@@ -46,7 +46,7 @@ export const commit = async (): Promise<void> => {
       latestInfo.launcherId.toString("hex")
     );
 
-    const datalayer = new DataIntegrityTree(storeId.toString('hex'), {
+    const datalayer = new DataIntegrityTree(storeId.toString("hex"), {
       storageMode: "local",
       storeDir: DIG_FOLDER_PATH,
       disableInitialize: true,
@@ -65,6 +65,10 @@ export const commit = async (): Promise<void> => {
 
     const newRootHash = datalayer.commit();
 
+    if (!newRootHash) {
+      return;
+    }
+
     const totalBytes = calculateFolderSize(DIG_FOLDER_PATH);
 
     console.log(
@@ -72,6 +76,7 @@ export const commit = async (): Promise<void> => {
     );
 
     const updatedStoreInfo = await updateDataStoreMetadata({
+      ...latestInfo.metadata,
       rootHash: Buffer.from(newRootHash, "hex"),
       bytes: totalBytes,
     });
@@ -89,6 +94,13 @@ export const commit = async (): Promise<void> => {
     if (!storeIntegrityCheck) {
       throw new Error("Store integrity check failed.");
     }
+
+    await waitForPromise(
+      () => getLatestStoreInfo(storeId),
+      "Finalizing commit...",
+      "Commit successful",
+      "Failed to commit."
+    );
 
     console.log("Commit successful");
   } catch (error: any) {