+++ src/actions/init.ts
@@ -1,40 +1,36 @@
 import * as fs from "fs";
-import * as path from "path";
 import { createDataLayerStore } from "../blockchain/datalayer";
-import { DataIntegrityTree, DataIntegrityTreeOptions } from "../DataIntegrityTree";
-import { DIG_FOLDER_PATH, MIN_HEIGHT, getHeightFilePath } from "../utils/config";
-import { askToDeleteAndReinit } from "../prompts";
+import {
+  DataIntegrityTree,
+  DataIntegrityTreeOptions,
+} from "../DataIntegrityTree";
+import {
+  DIG_FOLDER_PATH,
+  MIN_HEIGHT,
+  getHeightFilePath,
+  setActiveStore,
+  CONFIG_FILE_PATH,
+  createInitialConfig,
+} from "../utils/config";
 import { CreateStoreUserInputs } from "../types";
 import { getPeer } from "../blockchain/peer";
+import { getLatestStoreInfo } from "../blockchain/datastore";
+import { waitForPromise } from "../utils";
 
-export const init = async (inputs: CreateStoreUserInputs = {}): Promise<void> => {
-  if (fs.existsSync(DIG_FOLDER_PATH)) {
-    const shouldDelete = await askToDeleteAndReinit();
-
-    if (!shouldDelete) {
-      console.log("Initialization aborted.");
-      return;
-    }
-
-    // Delete the .dig directory
-    fs.rmSync(DIG_FOLDER_PATH, { recursive: true, force: true });
-    console.log(".dig folder deleted.");
+export const init = async (
+  inputs: CreateStoreUserInputs = {}
+): Promise<void> => {
+  if (!fs.existsSync(DIG_FOLDER_PATH)) {
+    fs.mkdirSync(DIG_FOLDER_PATH);
   }
 
-  // Re-create the .dig directory
-  fs.mkdirSync(DIG_FOLDER_PATH);
-
-  if (!fs.existsSync(path.join(process.cwd(), 'dig.config.json'))) {
-    const initialConfig = { deploy_dir: "./dist", origin: "" };
-    fs.writeFileSync(path.join(process.cwd(), 'dig.config.json'), JSON.stringify(initialConfig, null, 4));
-    console.log("Created dig.config.json file.");
+  if (!fs.existsSync(CONFIG_FILE_PATH)) {
+    createInitialConfig();
   }
 
   const peer = await getPeer();
   const currentHeight = (await peer.getPeak()) || MIN_HEIGHT;
   const currentHeaderHash = await peer.getHeaderHash(currentHeight);
-
-  
 
   const storeInfo = await createDataLayerStore(inputs);
   if (storeInfo) {
@@ -43,7 +39,6 @@ export const init = async (inputs: CreateStoreUserInputs = {}): Promise<void> =>
     const options: DataIntegrityTreeOptions = {
       storageMode: "local",
       storeDir: DIG_FOLDER_PATH,
-      disableInitialize: true
     };
 
     new DataIntegrityTree(storeId, options);
@@ -56,7 +51,20 @@ export const init = async (inputs: CreateStoreUserInputs = {}): Promise<void> =>
       })
     );
 
-    console.log(`Store initialized at Block Height: ${currentHeight} | ${currentHeaderHash.toString('hex')}`);
+    setActiveStore(storeId);
+
+    await waitForPromise(
+      () => getLatestStoreInfo(Buffer.from(storeId, "hex")),
+      "Final store initialization...",
+      "Store initialized.",
+      "Failed to initialize the data layer store."
+    );
+
+    console.log(
+      `Store initialized at Block Height: ${currentHeight} | ${currentHeaderHash.toString(
+        "hex"
+      )}`
+    );
   } else {
     console.log("Failed to initialize the data layer store.");
     fs.rmSync(DIG_FOLDER_PATH, { recursive: true, force: true });