+++ src/content_server/index.ts
@@ -1,0 +1,3 @@
+import { startPreviewServer } from './server';
+
+startPreviewServer();