+++ src/propagation_server/server.ts
@@ -1,0 +1,17 @@
+import { server, PORT } from "./app";
+
+const startPreviewServer = (): Promise<void> => {
+  return new Promise((resolve, reject) => {
+    try {
+      server.listen(PORT, async () => {
+        console.log(`DIG Propagation Server Started on port ${PORT}`);
+      });
+
+      server.on("close", resolve);
+    } catch (error) {
+      reject(error);
+    }
+  });
+};
+
+export { startPreviewServer };