+++ src/yargs/commands.ts
@@ -1,6 +1,6 @@
 import yargs, { Argv } from "yargs";
 import { handlers } from "./handlers";
-import { CreateStoreUserInputs } from '../types';
+import { CreateStoreUserInputs } from "../types";
 
 export function initCommand(yargs: Argv<{}>) {
   return yargs.command<CreateStoreUserInputs>(
@@ -68,12 +68,20 @@ export function pullCommand(yargs: Argv<{}>) {
 }
 
 export function cloneCommand(yargs: Argv<{}>) {
-  return yargs.command(
-    "clone",
-    "Clone a data store",
-    {},
-    handlers.clone
-  );
+    // @ts-ignore
+    return yargs.command<{ storeId: string }>(
+      "clone <storeId>",
+      "Clones a datastore from a remote remote",
+      (yargs: Argv<{ storeId: string }>) => {
+        return yargs.positional("storeId", {
+          type: "string",
+          describe: "The storId to clone down",
+        });
+      },
+      async (argv: {storeId: string}) => {
+        await handlers.clone(argv.storeId);
+      }
+    );
 }
 
 export function storeCommand(yargs: Argv<{}>) {
@@ -134,16 +142,18 @@ export function storeCommand(yargs: Argv<{}>) {
 
 export function remoteCommand(yargs: Argv<{}>) {
   // @ts-ignore
-  return yargs.command<{ connectionString: string }>(
-    "remote set <originConnectionString>",
-    "Set a datastore remote origin",
-    (yargs: Argv<{ connectionString: string }>) => {
-      return yargs.positional("connectionString", {
+  return yargs.command<{ peer: string }>(
+    "remote set <peer>",
+    "Set a datastore remote remote",
+    (yargs: Argv<{ peer: string }>) => {
+      return yargs.positional("peer", {
         type: "string",
-        describe: "The connection string for the datastore remote origin",
+        describe: "The host of the peer to set as the remote remote",
       });
     },
-    async (argv: {connectionString: string}) => await handlers.setRemote(argv.connectionString)
+    async (argv: {peer: string}) => {
+      await handlers.setRemote(argv.peer);
+    }
   );
 }
 
@@ -161,9 +171,10 @@ export function keysCommand(yargs: Argv<{}>) {
         })
         .option("mnemonic", {
           type: "string",
-          describe: "Mnemonic seed phrase for import (only for 'import' action)",
+          describe:
+            "Mnemonic seed phrase for import (only for 'import' action)",
         })
-        .strict();  // Ensures that only the defined options are accepted
+        .strict(); // Ensures that only the defined options are accepted
     },
     async (argv: { action: string; mnemonic?: string }) => {
       await handlers.manageKeys(argv.action, argv.mnemonic);
@@ -191,9 +202,9 @@ export function loginCommand(yargs: Argv<{}>) {
             throw new Error("--user and --pass must be provided together");
           }
           return true;
-        })
+        });
     },
-    async (argv: {user: string, pass: string})=> {
+    async (argv: { user: string; pass: string }) => {
       await handlers.login(argv.user, argv.pass);
     }
   );